Fix: uint8_t must be printed as integer, not single character

+++ lilac-core/src/hierarchy.cxx
@@ -20,7 +20,7 @@ namespace lilac::core
         ss << indents << Kind << ',' << Name << ',' << ActualName;
         if (Kind == HOK_Constant)
         {
-            ss << ',' << Constant.Signed << ',' << Constant.Size << ',';
+            ss << ',' << Constant.Signed << ',' << static_cast<int>(Constant.Size) << ',';
             if (Constant.Signed)
                 ss << Constant.ConstantSigned;
             else
